+++ src/models/subscription.rs
@@ -31,11 +31,11 @@ pub struct Subscription {
     #[serde(rename = "period")]
     pub period: models::SubscriptionPeriod,
     #[serde(rename = "tier")]
-    pub tier: f64,
+    pub tier: i32,
 }
 
 impl Subscription {
-    pub fn new(id: String, steam_item_id: String, amount: f64, description: String, period: models::SubscriptionPeriod, tier: f64) -> Subscription {
+    pub fn new(id: String, steam_item_id: String, amount: f64, description: String, period: models::SubscriptionPeriod, tier: i32) -> Subscription {
         Subscription {
             id,
             steam_item_id,
